+++ pkg/worker/k8srunner.go
@@ -9,6 +9,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/iver-wharf/wharf-cmd/pkg/resultstore"
 	"github.com/iver-wharf/wharf-cmd/pkg/tarutil"
 	"github.com/iver-wharf/wharf-cmd/pkg/wharfyml"
 	"github.com/iver-wharf/wharf-cmd/pkg/worker/workermodel"
@@ -28,8 +29,8 @@ var podInitContinueArgs = []string{"killall", "-s", "SIGINT", "sleep"}
 
 // NewK8s is a helper function that creates a new builder using the
 // NewK8sStepRunnerFactory.
-func NewK8s(ctx context.Context, def wharfyml.Definition, namespace string, restConfig *rest.Config, opts BuildOptions) (Builder, error) {
-	stageFactory, err := NewK8sStageRunnerFactory(namespace, restConfig)
+func NewK8s(ctx context.Context, def wharfyml.Definition, namespace string, restConfig *rest.Config, store resultstore.Store, opts BuildOptions) (Builder, error) {
+	stageFactory, err := NewK8sStageRunnerFactory(namespace, restConfig, store)
 	if err != nil {
 		return nil, err
 	}
@@ -38,8 +39,8 @@ func NewK8s(ctx context.Context, def wharfyml.Definition, namespace string, rest
 
 // NewK8sStageRunnerFactory is a helper function that creates a new stage runner
 // factory using the NewK8sStepRunnerFactory.
-func NewK8sStageRunnerFactory(namespace string, restConfig *rest.Config) (StageRunnerFactory, error) {
-	stepFactory, err := NewK8sStepRunnerFactory(namespace, restConfig)
+func NewK8sStageRunnerFactory(namespace string, restConfig *rest.Config, store resultstore.Store) (StageRunnerFactory, error) {
+	stepFactory, err := NewK8sStepRunnerFactory(namespace, restConfig, store)
 	if err != nil {
 		return nil, err
 	}
@@ -49,7 +50,7 @@ func NewK8sStageRunnerFactory(namespace string, restConfig *rest.Config) (StageR
 // NewK8sStepRunnerFactory returns a new step runner factory that creates
 // step runners with implementation that targets Kubernetes using a specific
 // Kubernetes namespace and REST config.
-func NewK8sStepRunnerFactory(namespace string, restConfig *rest.Config) (StepRunnerFactory, error) {
+func NewK8sStepRunnerFactory(namespace string, restConfig *rest.Config, store resultstore.Store) (StepRunnerFactory, error) {
 	clientset, err := kubernetes.NewForConfig(restConfig)
 	if err != nil {
 		return nil, err
@@ -58,6 +59,7 @@ func NewK8sStepRunnerFactory(namespace string, restConfig *rest.Config) (StepRun
 		namespace:  namespace,
 		restConfig: restConfig,
 		clientset:  clientset,
+		store:      store,
 	}, nil
 }
 
@@ -65,10 +67,11 @@ type k8sStepRunnerFactory struct {
 	namespace  string
 	restConfig *rest.Config
 	clientset  *kubernetes.Clientset
+	store      resultstore.Store
 }
 
 func (f k8sStepRunnerFactory) NewStepRunner(
-	ctx context.Context, step wharfyml.Step) (StepRunner, error) {
+	ctx context.Context, step wharfyml.Step, stepID uint64) (StepRunner, error) {
 	ctx = contextWithStepName(ctx, step.Name)
 	pod, err := getPodSpec(ctx, step)
 	if err != nil {
@@ -82,6 +85,8 @@ func (f k8sStepRunnerFactory) NewStepRunner(
 		restConfig: f.restConfig,
 		clientset:  f.clientset,
 		pods:       f.clientset.CoreV1().Pods(f.namespace),
+		store:      f.store,
+		stepID:     stepID,
 	}
 	if err := r.dryRunStepError(ctx); err != nil {
 		return nil, fmt.Errorf("dry-run: %w", err)
@@ -97,6 +102,8 @@ type k8sStepRunner struct {
 	restConfig *rest.Config
 	clientset  *kubernetes.Clientset
 	pods       corev1.PodInterface
+	store      resultstore.Store
+	stepID     uint64
 }
 
 func (r k8sStepRunner) Step() wharfyml.Step {
@@ -113,6 +120,7 @@ func (r k8sStepRunner) RunStep(ctx context.Context) StepResult {
 	} else if err != nil {
 		status = workermodel.StatusFailed
 	}
+	r.addStatusUpdate(status)
 	return StepResult{
 		Name:     r.step.Name,
 		Status:   status,
@@ -145,6 +153,7 @@ func (r k8sStepRunner) runStepError(ctx context.Context) error {
 		WithString("step", r.step.Name).
 		WithString("pod", r.pod.GenerateName).
 		Message("Creating pod.")
+	r.addStatusUpdate(workermodel.StatusInitializing)
 	newPod, err := r.pods.Create(ctx, r.pod, metav1.CreateOptions{})
 	if err != nil {
 		return fmt.Errorf("create pod: %w", err)
@@ -161,7 +170,6 @@ func (r k8sStepRunner) runStepError(ctx context.Context) error {
 	if err := r.waitForInitContainerRunning(ctx, newPod.ObjectMeta); err != nil {
 		return fmt.Errorf("wait for init container: %w", err)
 	}
-
 	log.Debug().WithFunc(logFunc).Message("Transferring repo to init container.")
 	if err := r.copyDirToPod(ctx, ".", "/mnt/repo", r.namespace, newPod.Name, "init"); err != nil {
 		return fmt.Errorf("transfer repo: %w", err)
@@ -170,6 +178,7 @@ func (r k8sStepRunner) runStepError(ctx context.Context) error {
 	if err := r.continueInitContainer(newPod.Name); err != nil {
 		return fmt.Errorf("continue init container: %w", err)
 	}
+	r.addStatusUpdate(workermodel.StatusRunning)
 	log.Debug().WithFunc(logFunc).Message("Waiting for app container to start.")
 	if err := r.waitForAppContainerRunningOrDone(ctx, newPod.ObjectMeta); err != nil {
 		if err := r.readLogs(ctx, newPod.Name, &v1.PodLogOptions{}); err != nil {
@@ -180,7 +189,7 @@ func (r k8sStepRunner) runStepError(ctx context.Context) error {
 	}
 
 	log.Debug().WithFunc(logFunc).Message("App container running. Streaming logs.")
-	if err := r.readLogs(ctx, newPod.Name, &v1.PodLogOptions{Follow: true}); err != nil {
+	if err := r.readLogs(ctx, newPod.Name, &v1.PodLogOptions{Follow: true, Timestamps: true}); err != nil {
 		return fmt.Errorf("stream logs: %w", err)
 	}
 	log.Debug().WithFunc(logFunc).Message("Logs ended. Waiting for termination.")
@@ -264,6 +273,10 @@ func (r k8sStepRunner) readLogs(ctx context.Context, podName string, opts *v1.Po
 	}
 	defer readCloser.Close()
 	scanner := bufio.NewScanner(readCloser)
+	writer, err := r.store.OpenLogWriter(uint64(r.stepID))
+	if err != nil {
+		r.log.Error().WithError(err).Message("Failed to open log writer. No logs will be written.")
+	}
 	for scanner.Scan() {
 		txt := scanner.Text()
 		idx := strings.LastIndexByte(txt, '\r')
@@ -271,6 +284,11 @@ func (r k8sStepRunner) readLogs(ctx context.Context, podName string, opts *v1.Po
 			txt = txt[idx+1:]
 		}
 		r.log.Info().Message(txt)
+		if writer != nil {
+			if err := writer.WriteLogLine(txt); err != nil {
+				r.log.Error().WithError(err).Message("Failed to write log line.")
+			}
+		}
 	}
 	return scanner.Err()
 }
@@ -363,4 +381,8 @@ func execInPod(c *rest.Config, namespace, podName string, execOpts *v1.PodExecOp
 		SubResource("exec").
 		VersionedParams(execOpts, scheme.ParameterCodec)
 	return remotecommand.NewSPDYExecutor(c, "POST", req.URL())
+}
+
+func (r *k8sStepRunner) addStatusUpdate(status workermodel.Status) {
+	r.store.AddStatusUpdate(r.stepID, time.Now(), status)
 }